requirements.txt
PyPDF2>=3.0.1
python-docx>=1.1.0
reportlab>=4.2.2
requests>=2.32.3
tenacity>=8.2.3
sentence-transformers>=3.0.1
numpy>=1.26.4
pandas>=2.2.2
ollama>=0.3.0
transformers>=4.43.3
datasets>=2.20.0
accelerate>=0.33.0
peft>=0.13.2
safetensors>=0.4.3
